feat: Finished the first version of the app.
The entire site is designed to be static for now since I'm trying to
keep it small and fast while I have little content on it.

In the future, I might want to start making things a bit more dynamic
with a separate admin routes for uploading and updating projects and
maybe even blogs. That's something to keep in mind for the future.